refactor(repository): make mixin more adaptable and remove soft crud classes
changes the type of base class being accepted by mixin to a more explicit interface
removes all soft crud classes being exported that can be replaced with mixin

BREAKING CHANGE:
Removes SoftCrudRepository, SequelizeSoftCrudRepository and DefaultTransactionSoftCrudRepository in
favour of the provided mixin.

GH-140
