+++ package.json
@@ -9,20 +9,6 @@
   ],
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "exports": {
-    ".": "./dist/index.js",
-    "./sequelize": {
-      "types": "./dist/repositories/sequelize/index.d.ts",
-      "default": "./dist/repositories/sequelize/index.js"
-    }
-  },
-  "typesVersions": {
-    "*": {
-      "sequelize": [
-        "./dist/repositories/sequelize/index.d.ts"
-      ]
-    }
-  },
   "engines": {
     "node": ">=16"
   },
