+++ src/index.ts
@@ -2,5 +2,4 @@ export * from './component';
 export * from './error-keys';
 export * from './mixins';
 export * from './models';
-export * from './repositories';
 export * from './types';
