+++ src/mixins/index.ts
@@ -1,2 +1,2 @@
-export * from './soft-crud.repository.mixin';
 export * from './soft-delete-entity.mixin';
+export * from './soft-crud.repository.mixin';
