--- src/repositories/index.ts
@@ -1,2 +1,0 @@
-export * from './soft-crud.repository.base';
-export * from './default-transaction-soft-crud.repository.base';
