--- src/repositories/sequelize/index.ts
@@ -1,1 +1,0 @@
-export * from './sequelize.soft-crud.repository.base';
